requirements.txt
streamlit==1.29.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiosqlite==0.19.0
openai==1.3.5
python-dotenv==1.0.0
aiofiles==23.2.1
aiohttp==3.9.1
pydantic-settings==2.1.0
slack-sdk>=3.26.0
httpx>=0.25.0
redis>=5.0.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
alembic>=1.13.0
sse-starlette>=1.8.0
